Share the temporary project directory helper between tests
